server: fix debugging under GDB
Don't use a temporary mapping for copying, but pread(). In the crammed
address space in GDB, the temporary mapping would steal the memory
area meant for the destination map. There's no need to pwrite() the
area to memfd, it will be done later.

Use also mmap() flag MAP_FIXED_NOREPLACE to get a warning of such
collisions.

Preallocate the memory area. Previously it was assumed that the memory
areas next to the first mmap() would be free but this may not be true
when using GDB. This simplifies memory mapping sequence, so we can
drop the second pass.

Signed-off-by: Topi Miettinen <toiwoton@gmail.com>